// apps/web/generated/api/models/userMeResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Freelance Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserMeResponse {
  user_id: string;
  email: string;
  roles: string[];
  permissions: string[];
  freelancer_profile_id?: string | null;
  freelancer_onboarding_needed?: boolean;
  freelancer_approval_status?: string | null;
  freelancer_level?: string | null;
}